move menuItems into own function
