+++ src/functions/index.ts
@@ -8,3 +8,4 @@ export {
 } from "./folderFunctions";
 export { checkAndAddPremades, applyPremades } from "./premadesFunctions";
 export { collisionInfo } from "./collisions";
+export { createMenuItems } from "./createMenuItems";
